backend_final/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
email-validator==2.1.0.post1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pdfplumber==0.10.3
google-generativeai==0.4.1
pillow==10.1.0
qrcode[pil]==7.4.2
openai-whisper
langdetect==1.0.9
sentence-transformers==2.4.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.26.0
torch==2.2.0
torchaudio==2.2.0
numpy==1.26.4
PyMuPDF==1.24.10
beautifulsoup4==4.12.3
